% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barleyCIM}
\alias{barleyCIM}
\title{Pre-computed CIM output barley}
\format{
An object of class \code{QTLMPP} (inherits from \code{GWAS}) of length 5.
}
\usage{
barleyCIM
}
\description{
Pre-computed CIM output for barley data used in vignette.
}
\keyword{datasets}
