% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barleyMQM}
\alias{barleyMQM}
\title{Pre-computed MQM output barley}
\format{
An object of class \code{QTLMPP} (inherits from \code{GWAS}) of length 5.
}
\usage{
barleyMQM
}
\description{
Pre-computed MQM output for barley data used in vignette.
}
\keyword{datasets}
