% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barleyPheno}
\alias{barleyPheno}
\title{Phenotypic data for awn length in barley}
\format{
A data.frame with 916 rows and 3 columns.
\describe{
\item{genotype}{Genotype}
\item{Awn_length}{Awn_length in mm.}
\item{cross}{The cross the genotype belongs to.}
}
}
\usage{
barleyPheno
}
\description{
Phenotypic data for awn length in barley
}
\references{
Fine mapping of a major QTL for awn length in barley using a
multiparent mapping population. Liller CB, Walla A, Boer MP, Hedley P,
Macaulay M, Effgen S, von Korff M, van Esse GW, Koornneef M.
Theor Appl Genet. 2017 Feb;130(2):269-281. \doi{10.1007/s00122-016-2807-y}.
}
\keyword{datasets}
