% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.distance}
\alias{calc.distance}
\title{cofactor selection based on the QTL windowsize}
\usage{
calc.distance(markers, m1, m2)
}
\description{
cofactor selection based on the QTL windowsize
}
\keyword{internal}
