% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcDistance}
\alias{calcDistance}
\title{cofactor selection based on the QTL windowsize}
\usage{
calcDistance(map, m1, m2)
}
\description{
cofactor selection based on the QTL windowsize
}
\keyword{internal}
