% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBDmpp.R
\name{calcIBDmpp}
\alias{calcIBDmpp}
\title{IBDs calculation and prepare the design matrix}
\usage{
calcIBDmpp(
  crossNames,
  locFiles,
  quaDat,
  poptypes,
  mapFile,
  evaldist,
  verbose = FALSE
)
}
\arguments{
\item{crossNames}{A character vector, the names of the crosses.}

\item{locFiles}{A vector of locations of the loc files.}

\item{quaDat}{A list of one or more data.frames with phenotypic information.}

\item{poptypes}{A character vector with population types.}

\item{mapFile}{A vector with the location of the map file.}

\item{evaldist}{A numeric value, the maximum distance in cM between
evaluation points.}

\item{verbose}{Should progress be printed?}
}
\description{
IBDs calculation and prepare the design matrix
}
