% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBDmppOrig.R
\name{calcIBDmppOrig}
\alias{calcIBDmppOrig}
\title{IBDs calculation and prepare the design matrix}
\usage{
calcIBDmppOrig(
  par.names,
  cross.names,
  loc.names,
  qua.names,
  pop.types,
  mapfile,
  evaldist
)
}
\arguments{
\item{par.names}{...}

\item{cross.names}{...}

\item{loc.names}{...}

\item{qua.names}{...}

\item{pop.types}{...}

\item{mapfile}{...}

\item{evaldist}{...}
}
\description{
IBDs calculation and prepare the design matrix
}
