% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIBDtoMPP.R
\name{convertIBDtoMPP}
\alias{convertIBDtoMPP}
\title{Convert IBDprob to gDataMPP}
\usage{
convertIBDtoMPP(IBDprob, pheno)
}
\arguments{
\item{IBDprob}{An object of class IBDprob}

\item{pheno}{A data frame with at least columns "genotype" for the
"genotype" and one or more numerical columns containing phenotypic
information. A column "cross" can be used for indicating the cross the
genotype comes from. This column is ignored if the \code{IBDprob} has an
attribute genoCross containing this information (the default behaviour).}
}
\description{
Function for converting an object of class IBDprob (computed or imported
using statgenIBD) to an object of class gDataMPP.
}
