% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectPlot.R
\name{effectPlot}
\alias{effectPlot}
\title{Create a matrix plot}
\usage{
effectPlot(
  effectDat,
  signSnp,
  map,
  xLab = "Chromosomes",
  yLab = "Parents",
  chrBoundaries,
  title = NULL,
  trait = NULL,
  ...,
  output = TRUE
)
}
\description{
Create a matrix plot of effect sizes and directions per parent for
significant snps.
}
\keyword{internal}
