% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maizeCIM}
\alias{maizeCIM}
\title{Pre-computed CIM output}
\format{
An object of class \code{QTLMPP} (inherits from \code{GWAS}) of length 5.
}
\usage{
maizeCIM
}
\description{
Pre-computed CIM output for maize data used in vignette.
}
\keyword{datasets}
