% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maizeMQM}
\alias{maizeMQM}
\title{Pre-computed MQM output maize}
\format{
An object of class \code{QTLMPP} (inherits from \code{GWAS}) of length 5.
}
\usage{
maizeMQM
}
\description{
Pre-computed MQM output for maize data used in vignette.
}
\keyword{datasets}
