% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maizeSIM}
\alias{maizeSIM}
\title{Pre-computed SIM output}
\format{
An object of class \code{QTLMPP} (inherits from \code{GWAS}) of length 5.
}
\usage{
maizeSIM
}
\description{
Pre-computed SIM output for maize data used in vignette.
}
\keyword{datasets}
