% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maizeSQM}
\alias{maizeSQM}
\title{Pre-computed SQM output maize}
\format{
An object of class \code{QTLMPP} (inherits from \code{GWAS}) of length 5.
}
\usage{
maizeSQM
}
\description{
Pre-computed SQM output for maize data used in vignette.
}
\keyword{datasets}
