% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parCIsPlot.R
\name{parCIsPlot}
\alias{parCIsPlot}
\title{Create a plot with effect sizes}
\usage{
parCIsPlot(
  parCIsDat,
  xLab = "Effect",
  yLab = "Parents",
  title = NULL,
  trait = NULL,
  ...,
  output = TRUE
)
}
\description{
Create a plot of QTL effect sizes and standard errors for each parent for
selected snps.
}
\keyword{internal}
