% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGData.R
\name{plot.gData}
\alias{plot.gData}
\title{Plot function for the class \code{gData}}
\usage{
\method{plot}{gData}(
  x,
  ...,
  plotType = c("genMap", "allGeno", "pedigree"),
  title = NULL,
  output = TRUE
)
}
\arguments{
\item{x}{An object of class \code{gData}.}

\item{...}{Further arguments to be passed on to the actual plotting
functions.}

\item{plotType}{A character string indicating the type of plot to be made.
One of "genMap", "allGeno" or "pedigree".}

\item{title}{A character string, the title of the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a ggplot object is invisibly returned.}
}
\value{
A ggplot object is invisibly returned.
}
\description{
Creates a plot of an object of S3 class \code{gData}. The following types of
plot can be made:
\itemize{
\item{\code{genMap}}{ A plot of the genetic map.}
\item{\code{allGeno}}{ A plot showing for all genotypes the IBD
probabilities of the parent with the highest probability per marker.}
\item{\code{pedigree}}{ A plot showing the structure of the pedigree of
the population.}
}
\cr
See the respective sections for more details on the plots.
}
\section{genMap}{

A plot is made showing the lengths of the chromosomes and the position of
the markers that are present in the map. It is possible to highlight one
or more markers using the extra parameter \code{highlight}.
}

\section{allGeno}{

A plot is made showing all genotypes and markers. Each combinaton of
genotype and marker is colored according to the parent with the highest
probability. A darker color indicates a higher probability.
}

\section{pedigree}{

A plot is made showing the structure of the pedigree for the population in
the \code{gData} object.
}

