% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLMPP.R
\name{print.summary.QTLMPP}
\alias{print.summary.QTLMPP}
\title{Print summary of object of class \code{summary.QTLMPP}}
\usage{
\method{print}{summary.QTLMPP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.QTLMPP}.}

\item{...}{Not used.}
}
\description{
Print summary of object of class \code{summary.QTLMPP}
}
