% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomQTLmodel.R
\name{randomQTLmodel}
\alias{randomQTLmodel}
\title{one position fitted in the mixed model}
\usage{
randomQTLmodel(
  MPPobj,
  trait = "pheno",
  scanPos = 1,
  cofPos = NULL,
  NULLmodel = FALSE
)
}
\description{
one position fitted in the mixed model
}
\keyword{internal}
