% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomQTLmodelOrig.R
\name{randomQTLmodelOrig}
\alias{randomQTLmodelOrig}
\title{one position fitted in the mixed model}
\usage{
randomQTLmodelOrig(
  MPPobj,
  trait.name = "pheno",
  scan_pos = 1,
  cof_pos = NULL,
  NULLmodel = FALSE
)
}
\description{
one position fitted in the mixed model
}
\keyword{internal}
