% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRabbit.R
\name{readRABBIT}
\alias{readRABBIT}
\title{Read IBD probabilities}
\usage{
readRABBIT(infile, pheno = NULL)
}
\arguments{
\item{infile}{A character string, a link to a .csv file with IBD
probabilities.}

\item{pheno}{A data frame with at least columns "genotype" for the
"genotype", "cross" indicating the cross the genotype comes from and one
or more numerical columns containing phenotypic information.}
}
\value{
A \code{gData} object with map and markers corresponding to the
imported information in the imported .csv file.
}
\description{
Read a file with IBD probabilities computed by the RABBIT software package.
}
\examples{
## Read RABBIT data for barley.
genoFile <- system.file("extdata", "BarleyMP_magicReconstruct_Summary.zip",
                       package = "statgenMPP")
barleyMPMPP <- readRABBIT(unzip(genoFile, exdir = tempdir()))

}
\references{
Reference to RABBIT.....
}
