% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanQTL.R
\name{scanQTL}
\alias{scanQTL}
\title{one-round genome scan with specified cofactor(s)}
\usage{
scanQTL(
  modDat,
  map,
  markers,
  parents,
  QTLwindow = 10,
  cof = NULL,
  Usc = NULL,
  trait = NULL,
  maxIter = 100,
  se = FALSE,
  parallel = FALSE,
  verbose = FALSE,
  biparNotConnect
)
}
\description{
one-round genome scan with specified cofactor(s)
}
\keyword{internal}
