% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanQTLOrig.R
\name{scanQTLOrig}
\alias{scanQTLOrig}
\title{one-round genome scan with specified cofactor(s)}
\usage{
scanQTLOrig(MPPobj, QTLwindow = 10, cof = NULL, trait.name = "pheno")
}
\description{
one-round genome scan with specified cofactor(s)
}
\keyword{internal}
