% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selQTLmppOrig.R
\name{selQTLmppOrig}
\alias{selQTLmppOrig}
\title{multi-round genome scans and select cofactors}
\usage{
selQTLmppOrig(
  MPPobj,
  QTLwindow = 10,
  threshold = 3,
  trait.name = "pheno",
  CIM = TRUE
)
}
\arguments{
\item{MPPobj}{...}

\item{QTLwindow}{...}

\item{threshold}{...}

\item{trait.name}{...}

\item{CIM}{...}
}
\description{
multi-round genome scans and select cofactors
}
