% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLmpp.R
\name{summary.GWAS}
\alias{summary.GWAS}
\title{Summary function for the class \code{QTLmpp}}
\usage{
\method{summary}{GWAS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{QTLmpp}.}

\item{...}{Not used.

#' @examples
\dontrun{
## Read phenotypic data.
pheno <- read.delim(system.file("extdata/multipop", "AxBxCpheno.txt",
                               package = "statgenMPP"))
## Rename first column to genotype.
colnames(pheno)[1] <- "genotype"#'

## Compute IBD probabilities for simulated population - AxB, AxC.
ABC <- calcIBDmpp(crossNames = c("AxB", "AxC"),
                  markerFiles = c(system.file("extdata/multipop", "AxB.txt",
                                              package = "statgenMPP"),
                                  system.file("extdata/multipop", "AxC.txt",
                                              package = "statgenMPP")),
                  pheno = pheno,
                  popType = "F4DH",
                  mapFile = system.file("extdata/multipop", "mapfile.txt",
                                        package = "statgenMPP"),
                  evalDist = 5)

## Composite Interval Mapping.
ABC_CIM <- selQTLmpp(ABC, trait = "pheno")

## Print summary.
summary(ABC_CIM)
}}
}
\description{
Gives a summary for an object of S3 class \code{QTLmpp}.
}
