% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBDmpp.R
\name{calcIBDmpp}
\alias{calcIBDmpp}
\title{IBDs calculation and prepare the design matrix}
\usage{
calcIBDmpp(
  crossNames,
  locFiles,
  quaDat,
  poptypes,
  mapFile,
  evaldist,
  verbose = FALSE
)
}
\arguments{
\item{crossNames}{...}

\item{locFiles}{...}

\item{quaDat}{...}

\item{poptypes}{...}

\item{mapFile}{...}

\item{evaldist}{...}
}
\description{
IBDs calculation and prepare the design matrix
}
