% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLmpp.R
\name{plot.QTLmpp}
\alias{plot.QTLmpp}
\title{Plot function for the class \code{QTLmpp}}
\usage{
\method{plot}{QTLmpp}(
  x,
  ...,
  plotType = c("manhattan", "parEffs", "QTLRegion"),
  title = NULL,
  output = TRUE
)
}
\arguments{
\item{x}{An object of class \code{QTLmpp}.}

\item{...}{further arguments to be passed on to the actual plotting
functions.}

\item{plotType}{A character string indicating the type of plot to be made.
One of "manhattan", and "matrix".}

\item{title}{A character string, the title of the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a list of ggplot objects is invisibly returned.}
}
\description{
Creates a plot of an object of S3 class \code{QTLmpp}. The following types of
plot can be made:
\itemize{
\item{a manhattan plot, i.e. a plot of LOD-scores per SNP}
\item{a plot of effect sizes and directions per parent}
}
Manhattan plots, qq plots and matrix plots are made for a single trait See
details for a detailed description of the plots and the plot options
specific to the different plots.
}
\section{Manhattan Plot}{

A LOD-profile of all marker positions and corresponding LOD-scores is
plotted. Significant markers are highlighted with red dots. By default these
are taken from the result of the GWAS analysis however the LOD-threshold for
significant parameters may be modified using the parameter \code{yThr}. The
treshold is plotted as a horizontal line. If there are previously known
marker effect, false positives and true negatives can also be marked.\cr
Extra parameter options:
\describe{
\item{\code{xLab}}{A character string, the x-axis label. Default =
\code{"Chromosomes"}}
\item{\code{yLab}}{A character string, the y-axis label. Default =
\code{-log10(p)}}
\item{\code{effects}}{A character vector, indicating which SNPs correspond
to a real (known) effect. Used for determining true/false positives and
false negatives. True positives are colored green, false positives orange and
false negatives yellow.}
\item{\code{colPalette}}{A color palette used for plotting. Default
coloring is done by chromosome, using black and grey.}
\item{\code{yThr}}{A numerical value for the LOD-threshold. The value from
the GWAS analysis is used as default.}
\item{\code{signLwd}}{A numerical value giving the thickness of the
points that are false/true positives/negatives. Default = 0.6}
\item{\code{lod}}{A positive numerical value. For the SNPs with a LOD-value
below this value, only 5\\% is plotted. The chance of a SNP being plotting is
proportional to its LOD-score. This option can be useful when plotting a
large number of SNPs.}
\item{\code{chr}}{A vector of chromosomes to be plotted. By default, all
chromosomes are plotted. Using this option allows restricting the plot to a
subset of chromosomes.}
}
}

\section{Parental effects Plot}{

A plot of effect sizes for each of the parents for the significant SNPs
found is created.\cr
Extra parameter options:
\describe{
\item{\code{xLab}}{A character string, the x-axis label. Default =
\code{"Chromosomes"}}
\item{\code{yLab}}{A character string, the y-axis label. Default =
\code{"Parents"}}
}
}

