% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomQTLmodel.R
\name{randomQTLmodel}
\alias{randomQTLmodel}
\title{one position fitted in the mixed model}
\usage{
randomQTLmodel(
  modDat,
  map,
  parents,
  trait = "pheno",
  scanMrk = NULL,
  cofMrk = NULL,
  NULLmodel = FALSE
)
}
\description{
one position fitted in the mixed model
}
\keyword{internal}
