% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRabbit.R
\name{readRABBIT}
\alias{readRABBIT}
\title{Read IBD probabilities}
\usage{
readRABBIT(infile)
}
\arguments{
\item{infile}{A character string, a link to a .csv file with IBD
probabilities.}
}
\value{
A \code{gData} object with map and markers corresponding to the
imported information in the imported .csv file.
}
\description{
Read a file with IBD probabilities computed by the RABBIT software package.
}
