% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRabbit.R
\name{readRABBIT}
\alias{readRABBIT}
\title{Read IBD probabilities}
\usage{
readRABBIT(infile, pheno = NULL, covar = NULL)
}
\arguments{
\item{infile}{A character string, a link to a .csv file with IBD
probabilities.}

\item{pheno}{A data.frame or a list of data.frames with phenotypic data,
with genotypes in the first column \code{genotype} and traits in the
following columns. The trait columns should be numerical columns only.
A list of data.frames can be used for replications, i.e. different
environments.}

\item{covar}{A data.frame with extra covariates per genotype. Genotypes
should be in the rows.}
}
\value{
A \code{gData} object with map and markers corresponding to the
imported information in the imported .csv file.
}
\description{
Read a file with IBD probabilities computed by the RABBIT software package.
}
\examples{
genoFile <- system.file("extdata", "BarleyMP_magicReconstruct_Summary.zip",
                       package = "statgenMPP")
barleyMPMPP <- readRABBIT(unzip(genoFile, exdir = tempdir()))

}
\references{
Fine mapping of a major QTL for awn length in barley using a
multiparent mapping population. Liller CB, Walla A, Boer MP, Hedley P,
Macaulay M, Effgen S, von Korff M, van Esse GW, Koornneef M.
Theor Appl Genet. 2017 Feb;130(2):269-281. doi: 10.1007/s00122-016-2807-y.
Epub 2016 Oct 12. PMID: 27734096
}
