% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanQTL.R
\name{scanQTL}
\alias{scanQTL}
\title{one-round genome scan with specified cofactor(s)}
\usage{
scanQTL(MPPobj, QTLwindow = 10, cof = NULL, trait.name = "pheno")
}
\description{
one-round genome scan with specified cofactor(s)
}
\keyword{internal}
