% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanQTL.R
\name{scanQTL}
\alias{scanQTL}
\title{one-round genome scan with specified cofactor(s)}
\usage{
scanQTL(
  modDat,
  map,
  parents,
  QTLwindow = 10,
  cof = NULL,
  trait = NULL,
  maxIter = 100,
  verbose = FALSE
)
}
\description{
one-round genome scan with specified cofactor(s)
}
\keyword{internal}
