% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selQTLmpp.R
\name{selQTLmpp}
\alias{selQTLmpp}
\title{multi-round genome scans and select cofactors}
\usage{
selQTLmpp(
  MPPobj,
  QTLwindow = 10,
  threshold = 3,
  trait.name = "pheno",
  CIM = TRUE
)
}
\description{
multi-round genome scans and select cofactors
}
