% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selQTLmpp.R
\name{selQTLmpp}
\alias{selQTLmpp}
\title{multi-round genome scans and select cofactors}
\usage{
selQTLmpp(
  MPPobj,
  trait = NULL,
  QTLwindow = 10,
  threshold = 3,
  CIM = TRUE,
  maxIter = 100,
  maxCofactors = 5,
  verbose = FALSE
)
}
\arguments{
\item{MPPobj}{...}

\item{trait}{...}

\item{QTLwindow}{...}

\item{threshold}{...}

\item{CIM}{...}

\item{maxIter}{...}

\item{maxCofactors}{...}

\item{verbose}{...}
}
\description{
multi-round genome scans and select cofactors
}
