% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenMPP.R
\docType{package}
\name{statgenMPP-package}
\alias{statgenMPP}
\alias{statgenMPP-package}
\title{statgenMPP: QTL mapping for Multi Parental Populations}
\description{
QTL mapping for Multi Parental Populations.
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Wenhao Li \email{wenhao.li@wur.nl} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

Other contributors:
\itemize{
  \item Martin Boer \email{martin.boer@wur.nl} [contributor]
}

}
\keyword{internal}
