% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenMPP.R
\docType{package}
\name{statgenMPP-package}
\alias{statgenMPP}
\alias{statgenMPP-package}
\title{statgenMPP: QTL Mapping for Multi Parent Populations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

For Multi Parent Populations (MPP) Identity By Descend (IBD) probabilities are computed using Hidden Markov Models. These probabilities are then used in a mixed model approach for QTL Mapping as described in Li et al. (\doi{10.1007/s00122-021-03919-7}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometris.github.io/statgenMPP/index.html}
  \item \url{https://github.com/Biometris/statgenMPP/}
  \item Report bugs at \url{https://github.com/Biometris/statgenMPP/issues}
}

}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Martin Boer \email{martin.boer@wur.nl} (\href{https://orcid.org/0000-0002-1879-4588}{ORCID})
  \item Wenhao Li (\href{https://orcid.org/0000-0001-5719-5775}{ORCID})
}

Other contributors:
\itemize{
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID}) [contributor]
}

}
\keyword{internal}
